Fix errors when running devserver. Make Place Entity constructor
